+++ lib/formatters/currency_input_formatter.dart
@@ -351,7 +351,8 @@ class CurrencyInputFormatter extends TextInputFormatter {
         var nextChar = '';
         if (caretPosition < newText.length - 1) {
           nextChar = newText[caretPosition];
-          if (!isDigit(nextChar, positiveOnly: true) || int.tryParse(nextChar) == 0) {
+          if (!isDigit(nextChar, positiveOnly: true) ||
+              int.tryParse(nextChar) == 0) {
             return true;
           }
         }